Adjust constants in rakefile
This PR renames:
* protoc_binary/protoc_version -> protoc_legacy_binary/protoc_legacy_version
* protoc_binary_2/protoc_version_2 -> protoc_binary/protoc_version

For background, protobuf has two major versions: 2 and 3.

The suffix for protoc_binary_2 implies that a 2.x series protobuf compiler is being used. But actually, protoc_binary_2 points to a 22.04 release, which is fairly new.

Contrast with protoc_binary which is pointing to protoc 3.5.1 which has been deprecated for several years (it's not even listed as deprecated in the public list of [releases](https://protobuf.dev/support/version-support/))

This PR also initializes protoc_binary and protoc_legacy_binary to their ultimate values from the get go. Compare this to before where they were both being set to `protoc` and then getting updated in setup_proto


This PR is purely a refactor, and does not change any generated code (this is proved by CI). After this PR, we should excise the legacy compiler, but that can be done in a separate PR.

Make it clearer what's happening