[protobuf] remove logs java protobuf and upgrade logs go compiler to non-legacy
This generated java is no longer used, instead the logs java backend copies the protobuf file 
and regenerates using their own generator. 

There are no differences between the legacy and modern proto compiler for Go, so we can upgrade to the latest protobuf compiler.

This is thefinal step before we can remove the legacy protobuf compiler.