Introduce host & container tags reporting for USM/CNM (#400)
* connections: Add index for host tags

The resolution of host tags in the backend is racy and can lead different pipelines (like CNM, USM) to
fail resolving the host tags, and by that to miss the tags which are important for the customer.

The agent supports reporting the host tags for a periodic interval, and we're importing this
logic to CNM and USM

* connections: Add index for local container tags

The resolution of container tags in the backend is racy and can lead different pipelines (like CNM, USM) to
fail resolving the tags, and by that to miss the tags which are important for the customer.