encoders: Introduce v3 of the encoder
The new version improves v2 by better handling duplications of tag-sets. V2 generates new index
for every time Encode method is being called. So calling Encode(Set1) and then again Encode(Set1)
yields two different indexes, and adds redundant bytes to the final buffer.
Also, any permutation of the given input, would result in a new index, although the order does not
matter.
V3 extends V2 by sorting and deduplicating the given input, and then ensures the same input
results in the same index. Thus achieving tight encoded buffer.
To have this functionality, we sarcifice the performance of V3 encoder, which is 50% slower
compared to V2. However, the final buffer is smaller (dependes on the duplications), but can
reach to 90% compression in some scenarios.

The new encoder allows us to spare redundant bytes in the payloads sent to the backend,
and for the backend, to reuse the outcome of getTags, if the index was already seen and analyzed.